requirements.txt
# ==================== CORE ====================
python-dateutil>=2.8.0
python-Levenshtein>=0.21.0

# ==================== DATA ====================
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# ==================== PDF/OCR ====================
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
PyPDF2>=3.0.0

# ==================== FUZZY MATCHING ====================
fuzzywuzzy>=0.18.0
rapidfuzz>=3.0.0

# ==================== WEB FRAMEWORK - FASTAPI ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# ==================== TESTING ====================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# ==================== DEVELOPMENT ====================
black>=23.7.0
flake8>=6.1.0
